Rename DisplayCard component to ItemCard to match its exported function name
