+++ client/src/pages/HomePage.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import NavigationBar from "../components/NavigationBar";
 import { useState, useEffect } from "react";
 import axios from "axios";
-import ItemCard from "../components/DisplayCard";
+import ItemCard from "../components/ItemCard";
 import { Buffer } from 'buffer';
 import { BACKEND_URL } from "../database/const";
 import { Box } from "@mui/material";
